fix: modify workqueue init
